+++ nebuchadnezzar/docs/changes.rst
@@ -11,6 +11,9 @@ Change Log
   it is determined that the user's version is older than the last
   release.
   See https://github.com/Connexions/nebuchadnezzar/issues/16
+- Resource/image files will no longer be copied into the working
+  directory when using ``neb get``.
+  See https://github.com/Connexions/nebuchadnezzar/issues/61
 
 4.0.0
 -----